+++ src/ts/infrastructure/storage/Storage.ts
@@ -39,6 +39,8 @@ enum StorageData {
   key = 'Test Daten',
   OnboardingAbgeschlossen = 'Onboarding abgeschlossen',
   OnboardingPersSnapshot = 'Onboarding Pers-Snapshot',
+  formularVersionCache = 'Formular-Version-Cache',
+  vorlagenPdfCache = 'Vorlagen-PDF-Cache',
 }
 
 export type TStorageData = keyof typeof StorageData;